Pricing and PIG value getters in Shape and Fixture (but values not balanced yet) All pricing and value structs moved into ObjectType enum AddShape and AddFixture moved to zone newZonePrice() added to PlayArea Current pointValue now stored in Shape and only calculated when changing getPassiveRate() added to Zone

+++ IncrementalGame/CurrencyFormat.swift
@@ -12,7 +12,7 @@ extension Int {
         let symbols = ["","K","M","B","T","q","Q"]
         
         var val = Double(self)
-        var suffix: Double
+        var suffix: Int
         
         switch self {
         case 0...999:
@@ -37,7 +37,7 @@ extension Int {
             suffix = 6
         }
         
-        let display = String(format: "%.4g ", val) + symbols[Int(suffix)]
+        let display = String(format: "%.4g ", val) + symbols[suffix]
         
         return display
     }
